+++ R/airsis_createMonitorObject.R
@@ -112,7 +112,7 @@ airsis_createMonitorObject <- function(startdate=strftime(lubridate::now(),"%Y01
   
   # See if anything gets through QC
   if ( nrow(tbl) == 0 ) {
-    logger.warn("No data remaining after QC")
+    logger.warn("No data remaining after QC") # This is more of a warning than some error in the data.
     stop("No data remaining after QC")
   }